requirements.txt
pandas
numpy
datetime
geopandas
azure-storage-blob